% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fromage}
\alias{fromage}
\title{French cheese data}
\format{
A data frame containing all the colours used in the palette:
\itemize{
\item nom: The name of the cheese
\item nom_alt: Alternative name(s) for the cheese, if any
\item departements: The department(s) where the cheese is typically produced; multiple departements are comma-delimited
\item fromage_famille: Type of cheese
\item lait_tratment: Treatment of milk to produce the cheese (raw, pasteurised, or either)
\item lait_animal: Type of milk (cow, sheep, goat, or other)
\item aoc: Whether the cheese is a "Appellation d'Origine Contrôlée"
\item aoc_year: The year that the cheese was given AOC status
}
}
\usage{
fromage
}
\description{
A dataset containing information about French cheeses.
}
\keyword{datasets}
