% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ref_fromagefam}
\alias{ref_fromagefam}
\title{Reference table for cheese families}
\format{
A data frame containing all the colours used in the palette:
\itemize{
\item fromage_famille: Family abbreviation, used in \code{fromage}
\item fromage_famille_fr: Family definition (FR)
\item fromage_famille_en: Family definition (EN)
}
}
\usage{
ref_fromagefam
}
\description{
Reference table between cheese family abbreviations and their definitions in French and English
}
\keyword{datasets}
