% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ref_region_departement}
\alias{ref_region_departement}
\title{Reference table between regions and departements}
\format{
A data frame containing all the colours used in the palette:
\itemize{
\item insee_code: INSEE departement code
\item region: Region name
\item departement: Departement name
}
}
\usage{
ref_region_departement
}
\description{
Reference table between French regions (admin-1) and departements (admin-2).
}
\keyword{datasets}
